Fix CreateFile emulation
The real CreateFile API only reads a DWORD of the disposition value. Therefore it is completely valid for malicious programs/shellcode to pass a QWORD (in x64) with invalid leading bytes and have the system API work (e.g. 0xcccccccc00000003 = OPEN_EXISTING). Without this change, Speakeasy fails to return a handle because it can't match the disposition parameter with any of the windef disposition values.